Update installer.sh
Corrected uplay variable for defining IP address
Commented out checks on interfaces so that it redeploys the file.  I made changes to the primary IP and wanted it to regenerate the interface file.  Not sure why you wouldn't want to just replace that file, as well as the others when the script runs every time.  Just thinking about future releases if you add more interfaces to lancache, they won't get updated when the user runs the installer again.